refactor: merge app metadata files into one file
